Rework & update Vendor-blobs
Noticed I missed 2 omx libs when updating to the Sharp 2.3.5 libs so to insure everything is updated as it should and we don't still have old Stock or other variant files mixed in or lingering around deleted all the files in the proprietary directory listed in the “triumph-vendor-blobs.mk”. I then rewrote each line for each listing to insure for example a line for an old Froyo MM Module not used in Gingerbread is not still mixed in with the lines to copy in the Gingerbread Modulus.

After using hex-ray with each individual omx lib learned that a decent majority of the libs list the Opencore libs found in all the different FB0 variants so have added these libs to insure the included omx libs have all the files dependencies needed to work properly. I have also been doing this with all the other files not just those used for omx so have also added in a few other files listed as dependencies for the other files we are adding in and not building from Source.

After looking over the device files for a multitude of similar devices found commits adding in lines to copy some files into /obj/ despite not receiving an error indicating the /obj/ is needed. So have added in a few of these lines selecting files half based on the above mentioned commits and my own experimentation.